# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scroll.io API - digital wellness tracker for reels/shorts
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  displayName: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  leetcodeUsername?: string | null;
  createdAt: string;
  totalReelsSeen?: number;
  totalMinutesWatched?: number;
  currentStreak?: number;
  bestStreak?: number;
  totalDaysBlocked?: number;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  leetcodeUsername?: string;
}

export interface PublicProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  totalReelsSeen?: number;
  totalMinutesWatched?: number;
  currentStreak?: number;
  bestStreak?: number;
  totalDaysBlocked?: number;
  rank?: number;
}

export interface DailyStats {
  date: string;
  reelsSeen: number;
  minutesWatched: number;
  instagramReels: number;
  youtubeShorts: number;
  appsBlocked?: boolean;
  unlockedByLeetcode?: boolean;
}

export interface StatsInput {
  date: string;
  reelsSeen: number;
  minutesWatched: number;
  instagramReels: number;
  youtubeShorts: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  username: string;
  displayName: string;
  value: number;
  currentStreak?: number;
}

export interface LeetcodeVerifyInput {
  leetcodeUsername: string;
}

export interface LeetcodeVerifyResponse {
  verified: boolean;
  solvedToday: number;
  unlockedApps: boolean;
  message?: string;
}

export interface LeetcodeStatus {
  verified: boolean;
  solvedToday: number;
  unlockedApps: boolean;
  requiredSolves: number;
  /** @nullable */
  leetcodeUsername?: string | null;
}

export interface BlockingStatus {
  instagramBlocked: boolean;
  youtubeBlocked: boolean;
  blockingEnabled: boolean;
  /** @nullable */
  unlockedUntil?: string | null;
}

export interface BlockingStatusInput {
  blockingEnabled: boolean;
  instagramBlocked?: boolean;
  youtubeBlocked?: boolean;
}

export type GetLeaderboardParams = {
metric?: GetLeaderboardMetric;
limit?: number;
};

export type GetLeaderboardMetric = typeof GetLeaderboardMetric[keyof typeof GetLeaderboardMetric];


export const GetLeaderboardMetric = {
  streak: 'streak',
  reelsBlocked: 'reelsBlocked',
  timeSaved: 'timeSaved',
} as const;