# lib/api-zod/src/generated/types/blockingStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scroll.io API - digital wellness tracker for reels/shorts
 * OpenAPI spec version: 0.1.0
 */

export interface BlockingStatus {
  instagramBlocked: boolean;
  youtubeBlocked: boolean;
  blockingEnabled: boolean;
  /** @nullable */
  unlockedUntil?: Date | null;
}