# lib/api-zod/src/generated/types/leetcodeVerifyResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scroll.io API - digital wellness tracker for reels/shorts
 * OpenAPI spec version: 0.1.0
 */

export interface LeetcodeVerifyResponse {
  verified: boolean;
  solvedToday: number;
  unlockedApps: boolean;
  message?: string;
}