# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scroll.io API - digital wellness tracker for reels/shorts
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  leetcodeUsername?: string | null;
  createdAt: Date;
  totalReelsSeen?: number;
  totalMinutesWatched?: number;
  currentStreak?: number;
  bestStreak?: number;
  totalDaysBlocked?: number;
}