# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scroll.io API - digital wellness tracker for reels/shorts
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  leetcodeUsername?: string;
}